app/requirements.txt
fastapi
uvicorn
pymavlink
pyserial
